requirements.txt
python-dotenv==1.0.1
requests==2.32.2
pytest==8.2.0
pytest-benchmark==4.0.0
faker==25.2.0
pytest-check==2.3.1          # “soft assertions” için
